[cryptofolio] Updating readme, moving refresh button
Updated `README.md` with new screenshots and listed all environment
variables that are used at the moment.

Moved the refresh button from the home page to the settings page since I
plan to make the refresh balace feature be implicit and the user
should't do that manually except he wants to.